pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "self-host-cloud-agent"
version = "0.1.0"
description = "The agent that completely manages your personal server"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [
    { name = "Avasis AI", email = "contact@avasis.ai" }
]
maintainers = [
    { name = "Avasis AI", email = "contact@avasis.ai" }
]
keywords = [
    "self-host",
    "homelab",
    "docker",
    "server-manager",
    "reverse-proxy",
    "ssl",
    "certbot",
    "traefik",
    "container-orchestration"
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    "docker>=6.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.urls]
Homepage = "https://github.com/avasis-ai/self-host-cloud-agent"
Documentation = "https://github.com/avasis-ai/self-host-cloud-agent#readme"
Repository = "https://github.com/avasis-ai/self-host-cloud-agent.git"
Issues = "https://github.com/avasis-ai/self-host-cloud-agent/issues"

[project.scripts]
self-host-cloud-agent = "self_host_cloud_agent.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/self_host_cloud_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"
multi_line_output = 3
